==> analysis/01_power_analysis.Rmd <==
---
title: "Power Analysis MP1 additional data collection"
author: "Manuel Bohn"
date: "6 Oct 2018"
output:
  html_notebook:
    code_folding: hide
    css: style.css
    theme: paper
    toc: yes
    toc_float: yes
---

```{r importing data, message=FALSE}
library(tidyverse)
library(lsr) # for cohen's d
library(pwr)
library(knitr)
library(broom)
```

```{r}
# create clade variable
species_data <- read.csv("../data/species_data.csv") %>% select(species, clade)
mp_data <- read.csv("../data/pilot_data/01_manyprimates_pilot_merged_data.csv") %>%
  left_join(species_data, by = 'species')
```

# Power analysis based on data pooled across conditions

## Effect sizes from Pilot

As a first step, we split the data by clade and get an overall effect sizes for comparison to chance across conditions. Data is aggregated per individual. Effect size is based on one sample t-test, chance level is 1/3.

```{r effect sizes from round 1 per clade, results = 'asis'}
effect_sizes <- mp_data %>%
  group_by(clade, subject_site) %>%
  summarise(correct = mean(correct)) %>%
  summarise(n = n(), mean = mean(correct), sd = sd(correct),
            t = t.test(correct, mu = 1/3)$statistic,
            p = t.test(correct, mu = 1/3)$p.value,
            d = cohensD(correct, mu = 1/3))

kable(effect_sizes, digits = 2)
```

## Power analysis 

Next we do the power analysis per clade based on the effect size from our Pilot. Desired power is .80, alpha is .05

```{r, results = 'asis'}
effect_sizes %>%
  group_by(clade) %>%
  summarise(N = pwr.t.test(d = d, sig.level = .05, power = .8, type = 'one')$n,
            sig.level = .05, power = .8) %>%
  kable(digits = 2)
```

- This suggests that the minimum number of subjects for any given ape species is 3. This is biased by the fact that we have only great apes, so we might want to consider a larger minimum sample size for gibbons.
- For Old World Monkeys, we need at least 6 subjects per species. 
- For New World Monkeys, we also need at least 6 subjects per species. 
- For Lemurs we need at least 4 subjects per species. The number is lower than in the case of monkeys because lemur performance was less variable during the pilot. 

***

# Power analysis based on delay conditions only

The long and medium delay conditions are more informative with respect to short-term memory abilities. Here is the same analysis as above based on data averaged across the two delay conditions.

## Effect sizes from Pilot

Again, effect size is based in one sample t-test, chance level is 1/3.

```{r effect sizes from round 1 per clade for delay, results = 'asis'}
effect_sizes_delay <- mp_data %>%
  filter(condition != 'short') %>%
  group_by(clade, subject_site) %>%
  summarise(correct = mean(correct)) %>%
  summarise(n = n(), mean = mean(correct), sd = sd(correct),
            t = t.test(correct, mu = 1/3)$statistic,
            p = t.test(correct, mu = 1/3)$p.value,
            d = cohensD(correct, mu = 1/3))

kable(effect_sizes_delay, digits = 2)
```

## Power analysis 

Again, desired power is .80, alpha is .05

```{r, results = 'asis'}
effect_sizes_delay %>%
  group_by(clade) %>%
  summarise(N = pwr.t.test(d = d, sig.level = .05, power = .8, type = 'one')$n,
            sig.level = .05, power = .8) %>%
  kable(digits = 2)
```

- Minimum number of subjects for apes based on delay conditions is 4.
- For Old World monkeys, we would need at least 9 subjects per species to achieve desired power. 
- For New World Monkeys, we would need at least 12 per species. 
- We would need at least 5 lemurs per species.
