==> analysis/03_phylogenetic_analysis.Rmd <==
---
title: "ManyPrimates1_Phylogenetic_Analysis"
output:
  html_notebook:
    code_folding: hide
    css: style.css
    theme: paper
    toc: yes
    toc_float: yes
---

```{r setup, include=FALSE}
library(ape)
library(tidyverse)
library(readxl)
library(googledrive)
library(brms)
library(broom)
library(ggthemes)
library(tidybayes)
library(ggridges)
library(ggpubr)
library(parallel)
```


# NOTE
Running the models takes a long time (12h on a HPC) - outputs are too large to put on GitHub (~7GB). If you want to access the model outputs directly, you have to download them via the following link. Code below expects files to be in `saves/`.

If the code doesn't work, you can download the files here: `https://keeper.mpdl.mpg.de/f/e78a9963b58f4a9190d8/?dl=1` but you will have to put them manually into the `saves` directory.

```{r}
# download zip file from Max Planck cloud storage
options(timeout=3600) # 1h, maybe you need more depending on your internet speed
download.file("https://keeper.mpdl.mpg.de/f/e78a9963b58f4a9190d8/?dl=1", 
              destfile = "../saves/saves.zip")

# unpack zip file to get .rds files
unzip("../saves/saves.zip", exdir = "../saves/")
```



# Data processing

## Read in STM data file

```{r}
data <- read_csv("../data/merged_data/ManyPrimates_mp1_merged_data.csv")
```

## Read in species level predictors and scale them.

Species `hamlyns_monkey` and `allens_swamp_monkey` will be excluded from the phylogenetic analysis because we did not find data for some predictors for those species. For `sichuan_snub_nosed_monkey` we actually don't have any experimental data.

```{r}
# drive_download("https://docs.google.com/spreadsheets/d/1RCLvbNWFph6bHvupetx7RDiAu5BWXK6PumgRo44ZoeM/edit#gid=1909432990",
#                path = "../data/species_predictors.xlsx", 
#                overwrite = T)

species_data <- read_xlsx("../data/species_predictors.xlsx", sheet = 1)%>% 
  filter(row_number() <= n()-3)%>% # remove last rows which contained a note
  mutate_if(is.character, tolower)%>%
  filter(species_english != "allens_swamp_monkey",
         species_english != "hamlyns_monkey",
         species_english != "sichuan_snub_nosed_monkey")%>%
  mutate(feeding_budget = as.numeric(feeding_budget),
         day_journey_length= as.numeric(day_journey_length),
         home_range_body_site_ratio = home_range/body_size)%>%
  mutate(resting_time_percent_activitybudget = scale(resting_time_percent_activitybudget, center = T, scale = T),
         home_range_body_site_ratio = scale(home_range_body_site_ratio, center = T, scale = T),
         home_range = scale(home_range, center = T, scale = T),
         feeding_budget = scale(feeding_budget, center = T, scale = T),
         dietary_breadth = scale(dietary_breadth, center = T, scale = T),
         vocal_repertoire = scale(vocal_repertoire, center = T, scale = T),
         group_size = scale(group_size, center = T, scale = T),
         day_journey_length = scale(day_journey_length, center = T, scale = T),
         percent_frugivory = scale(percent_frugivory, center = T, scale = T),
         feeding_budget = scale(feeding_budget, center = T, scale =T))

```

## Merge experimental data with species-level predictors

Exclude species for which we have no species-level data. `Gorilla` has to be re-coded to match the species name in the species-level predictor file.

```{r}
phylo_data <- data %>%
  rename(species_english = species)%>% # create a common column to merge by
  mutate(species_english = ifelse(species_english == "gorilla", "western_gorilla",species_english))%>%
  filter(species_english != "allens_swamp_monkey",
         species_english != "hamlyns_monkey",
         species_english != "sichuan_snub_nosed_monkey")%>%# re-code gorillas to match the species names 
  left_join(species_data)

# check if there are remaining NAs for the predictors

phylo_data %>%
  filter_at(vars(22:33),any_vars(is.na(.)))


```

## Recode the "short/long/medium" condition variable into medium-centred numeric delays:

```{r}
phylo_data <- phylo_data %>%
  mutate(delay = if_else(condition == "short", -1,
                         if_else(condition == "medium", 0, 1)),
         trial = scale(trial, center = T, scale = T))
```

# Phylogenetic tree

Tree-related processing:

 * Read phylogenetic tree from file
 * Check that all species labels in the data also occur as tip labels in the tree
 * Prune any species not in the data from the tree
 * Compute covariance matrix from tree under Brownian motion model

```{r}
tree <- read.tree("../phylo/mp_species.tree")
data_species <- phylo_data$species_english %>% unique
tree_species <- tree$tip.label
stopifnot(length(setdiff(data_species, intersect(tree_species, data_species))) == 0)
tree <- keep.tip(tree, data_species)
A <- vcv.phylo(tree)
```

# Phylogenetic models

```{r}
formulas <- list()
priors <- list()
```

## Model 0: Phylogenetic baseline model

```{r}
model_name <- "base_model"
formulas[[model_name]] <- correct ~ delay + (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(skew_normal(0, 1.5, 3), class=Intercept) +
                        prior(skew_normal(-1.4, 2, 3), coef=delay) +
                        prior(normal(0, 1), class=sd, coef=Intercept, group=subject_site) +
                        prior(normal(0, 1), class=sd, coef=Intercept, group=site) +
                        prior(normal(0, 1), class=sd, coef=Intercept, group=species_english) +
                        prior(normal(0, 0.5), class=sd, coef=delay, group=subject_site) +
                        prior(normal(0, 0.5), class=sd, coef=delay, group=site) +
                        prior(normal(0, 0.5), class=sd, coef=delay, group=species_english) +
                        prior(normal(0, 0.5), class=sd, coef=trial, group=subject_site) +
                        prior(normal(0, 0.5), class=sd, coef=trial, group=site) +
                        prior(normal(0, 0.5), class=sd, coef=trial, group=species_english) +
                        prior("lkj(1)", class = "cor")
```

## Model 1: Diurnal resting-time

Numeric predictor: `resting_time_percent_activitybudget`.

```{r}
model_name <- "diurnal_resting_model"
formulas[[model_name]] <- correct ~ delay * resting_time_percent_activitybudget +
                          (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 1), coef=resting_time_percent_activitybudget) +
                        prior(normal(0, 1), coef=`delay:resting_time_percent_activitybudget`)
```

## Model 2: Color vision

Categorical predictor: `color_vision`. 

```{r}
model_name <- "color_vision_model"
formulas[[model_name]] <- correct ~ delay * color_vision +
                          (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 2), coef=color_visionpolymorphic) +
                        prior(normal(0, 2), coef=color_visiontrichromatic) +
                        prior(normal(0, 2), coef=`delay:color_visionpolymorphic`) +
                        prior(normal(0, 2), coef=`delay:color_visiontrichromatic`)
```

## Model 3: Home_range/body size ratio, time spent feeding and range of consumed food items

numerical predictors: `home_range_body_site_ratio`, `feeding_budget` and `dietary_breadth`.

```{r}
model_name <- "home_range_feeding_budget_diet_breath_model"
formulas[[model_name]] <- correct ~ delay * home_range_body_site_ratio + delay * feeding_budget * dietary_breadth +
                          (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 1), coef=home_range_body_site_ratio) +
                        prior(normal(0, 1), coef=`delay:dietary_breadth`) +
                        prior(normal(0, 1), coef=`delay:home_range_body_site_ratio`) +
                        prior(normal(0, 1), coef=`delay:feeding_budget`) +
                        prior(normal(0, 1), coef=`delay:feeding_budget:dietary_breadth`) +
                        prior(normal(0, 1), coef=dietary_breadth) +
                        prior(normal(0, 1), coef=feeding_budget) +
                        prior(normal(0, 1), coef=`feeding_budget:dietary_breadth`)
```

## Model 3a: Home range

numerical predictor: `home_range`.

```{r}
model_name <- "home_range_model"
formulas[[model_name]] <- correct ~ delay * home_range +
                          (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 1), coef=home_range) +
                        prior(normal(0, 1), coef=`delay:home_range`)
```

## Model 3b: Feeding budget

numerical predictor: `feeding_budget`

```{r}
model_name <- "feeding_budget_model"
formulas[[model_name]] <- correct ~ delay * feeding_budget +
                          (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 1), coef=feeding_budget) +
                        prior(normal(0, 1), coef=`delay:feeding_budget`)
```

## Model 4: Vocal repertoire

Numeric predictor: `vocal_repertoire`. 

```{r}
model_name <- "vocal_repertoire_model"
formulas[[model_name]] <- correct ~ delay * vocal_repertoire +
                          (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 1), coef=vocal_repertoire) +
                        prior(normal(0, 1), coef=`delay:vocal_repertoire`)
```

## Model 5: Dietary breath

Numeric predictor: `dietary_breadth`. 

```{r}
model_name <- "dietary_breadth_model"
formulas[[model_name]] <- correct ~ delay * dietary_breadth +
                          (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 1), coef=dietary_breadth) +
                        prior(normal(0, 1), coef=`delay:dietary_breadth`)
```

## Model 6: Group size

Numeric predictor: `group_size`. 

```{r}
model_name <- "group_size_model"
formulas[[model_name]] <- correct ~ delay * group_size +
                          (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 1), coef=group_size) +
                        prior(normal(0, 1), coef=`delay:group_size`)
```

## Model 7: Day journey length and group size

Numeric predictors: `day_journey_length` and `group_size`. 

```{r}
model_name <- "day_journey_group_size_model"
formulas[[model_name]] <- correct ~ delay * day_journey_length * group_size +
                          (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 1), coef=group_size) +
                        prior(normal(0, 1), coef=day_journey_length) +
                        prior(normal(0, 1), coef=`day_journey_length:group_size`) +
                        prior(normal(0, 1), coef=`delay:group_size`) +
                        prior(normal(0, 1), coef=`delay:day_journey_length`) +
                        prior(normal(0, 1), coef=`delay:day_journey_length:group_size`)
```

## Model 7a: Day journey length

Numeric predictor: `day_journey_length`.

```{r}
model_name <- "day_journey_model"
formulas[[model_name]] <- correct ~ delay * day_journey_length +
                          (1 + delay + trial | subject_site + size + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 1), coef=day_journey_length) +
                        prior(normal(0, 1), coef=`delay:day_journey_length`)
```

## Model 8: Percent frugivory and terrestriality

Numeric predictor: `percent_frugivory`. Categorical predictor and `terrestriality`. 

```{r}
model_name <- "frugivory_terrestriality_model"
formulas[[model_name]] <- correct ~ delay * percent_frugivory * terrestriality +
                          (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 2), coef=terrestrialityterrestrial) +
                        prior(normal(0, 1), coef=percent_frugivory) +
                        prior(normal(0, 2), coef=`percent_frugivory:terrestrialityterrestrial`) +
                        prior(normal(0, 2), coef=`delay:terrestrialityterrestrial`) +
                        prior(normal(0, 1), coef=`delay:percent_frugivory`) +
                        prior(normal(0, 2), coef=`delay:percent_frugivory:terrestrialityterrestrial`)
```

## Model 8a: Percent frugivory

Numeric predictor: `percent_frugivory`.

```{r}
model_name <- "frugivory_model"
formulas[[model_name]] <- correct ~ delay * percent_frugivory +
                          (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 1), coef=percent_frugivory) +
                        prior(normal(0, 1), coef=`delay:percent_frugivory`)
```

## Model 8b: Terrestriality

Categorical predictor and `terrestriality`. 

```{r}
model_name <- "terrestriality_model"
formulas[[model_name]] <- correct ~ delay * terrestriality +
                          (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 2), coef=terrestrialityterrestrial) +
                        prior(normal(0, 2), coef=`delay:terrestrialityterrestrial`)
```

## Model 9: Diet diversity

Categorical predictor: `diet_diversity`. 

```{r}
model_name <- "diet_diversity_model"
formulas[[model_name]] <- correct ~ delay * diet_diversity +
                          (1 + delay + trial | subject_site + site + gr(species_english, cov=A))
priors[[model_name]] <- prior(normal(0, 2), coef=diet_diversityfolivore_frugivore) +
                        prior(normal(0, 2), coef=diet_diversityfrugivore) +
                        prior(normal(0, 2), coef=diet_diversitygummivore) +
                        prior(normal(0, 2), coef=diet_diversityinsectivore_frugivore) +
                        prior(normal(0, 2), coef=diet_diversityomnivore) +
                        prior(normal(0, 2), coef=`delay:diet_diversityfolivore_frugivore`) +
                        prior(normal(0, 2), coef=`delay:diet_diversityfrugivore`) +
                        prior(normal(0, 2), coef=`delay:diet_diversitygummivore`) +
                        prior(normal(0, 2), coef=`delay:diet_diversityinsectivore_frugivore`) +
                        prior(normal(0, 2), coef=`delay:diet_diversityomnivore`)
```

## Run all models in parallel
```{r}

# First, add the common base prior to all other priors
stopifnot(names(formulas)[1] == "base_model")
for(i in 2:length(priors)) {
  priors[[i]] <- priors[[1]] + priors[[i]]
}

# Derive filenames for model results from model names
filenames <- paste("../saves/phylo_", names(formulas), ".rds", sep="")

# Define a function to run, process and save a model given a formula and prior
run_model <- function(formula, prior, chains, iter, filename) {
  print(c(filename, formula, prior))
  model <- brm(formula, family=bernoulli(),
               data=phylo_data,
               data2 = list(A = A),
               prior=prior,
               chains=chains,
               iter=iter,
               cores=chains,
            )
  model <- add_criterion(model, "waic")
  saveRDS(model, filename)
}

# Run and save all models in parallel
chains <- 8
iter <- 5000
available_cores <- detectCores()
simultaneous_models <- floor(available_cores / chains)

## Turn warnings into errors so we won't get past the call to mcmapply if
## something goes wrong
options(warn=2)

mcmapply(run_model, formulas, priors, chains, iter, filenames,
         mc.cores=simultaneous_models)

## Restore default warning behaviour
options(warn=0)

# Load all model result and assign them nice variables names ready for the
# model comparison code below
for(i in 1:length(formulas)) {
  model <- readRDS(filenames[i])
  assign(names(formulas)[i], model)
}
```

# Model comparison

## Models from modelling challenge

```{r}
# Compute WAIC scores and SE
comp <-  loo_compare(
  base_model, 
  diurnal_resting_model,
  color_vision_model,
  home_range_feeding_budget_diet_breath_model,
  vocal_repertoire_model,
  dietary_breadth_model,
  group_size_model,
  day_journey_group_size_model,
  frugivory_terrestriality_model,
  diet_diversity_model,
  criterion = "waic")%>%
  as_tibble(rownames = "model")
  
# Compute WAIC weights
weights <- model_weights(
  base_model, 
  diurnal_resting_model,
  color_vision_model,
  home_range_feeding_budget_diet_breath_model,
  vocal_repertoire_model,
  dietary_breadth_model,
  group_size_model,
  day_journey_group_size_model,
  frugivory_terrestriality_model,
  diet_diversity_model, weights = "waic")%>%
  as_tibble() %>% 
  rename(weight = value) %>% 
  mutate(model  = c("base_model", 
                    "diurnal_resting_model",
                    "color_vision_model",
                    "home_range_feeding_budget_diet_breath_model",
                    "vocal_repertoire_model",
                    "dietary_breadth_model",
                    "group_size_model",
                    "day_journey_group_size_model",
                    "frugivory_terrestriality_model",
                    "diet_diversity_model"),
         weight = weight %>% round(digits = 2)) %>% 
  select(model, weight) 


# Combine weights and scores
model_comparison <- comp %>%
  select(model,waic, se_waic)%>%
  left_join(weights)%>%
  arrange(desc(weight))

#write_csv(model_comparison, "../saves/model_comparison.csv")

model_comparison <- read_csv("../saves/model_comparison.csv")

model_comparison
```
## Exploratory analysis:Include main effects models for all predictors

Some models were submitted with interaction terms between variables. The additional models below were built so that each predictor that was submitted as part of an interaction is also included in a model with only that predictor. 

```{r}
# Compute WAIC scores and SE
comp_exp <-  loo_compare(
  base_model,    # models from modelling challnege
  diurnal_resting_model,
  color_vision_model,
  home_range_feeding_budget_diet_breath_model,
  vocal_repertoire_model,
  dietary_breadth_model,
  group_size_model,
  day_journey_group_size_model,
  frugivory_terrestriality_model,
  diet_diversity_model,
  
  home_range_model,      # main effects models 
  feeding_budget_model,
  day_journey_model,
  frugivory_model,
  terrestriality_model,
  
  criterion = "waic")%>%
  as_tibble(rownames = "model")
  
# Compute WAIC weights
weights_exp <- model_weights(
  base_model, 
  diurnal_resting_model,
  color_vision_model,
  home_range_feeding_budget_diet_breath_model,
  vocal_repertoire_model,
  dietary_breadth_model,
  group_size_model,
  day_journey_group_size_model,
  frugivory_terrestriality_model,
  diet_diversity_model, 
  home_range_model,
  feeding_budget_model,
  day_journey_model,
  frugivory_model,
  terrestriality_model,
  weights = "waic")%>%
  as_tibble() %>% 
  rename(weight = value) %>% 
  mutate(model  = c("base_model", 
                    "diurnal_resting_model",
                    "color_vision_model",
                    "home_range_feeding_budget_diet_breath_model",
                    "vocal_repertoire_model",
                    "dietary_breadth_model",
                    "group_size_model",
                    "day_journey_group_size_model",
                    "frugivory_terrestriality_model",
                    "diet_diversity_model",
                    "home_range_model",
                    "feeding_budget_model",
                    "day_journey_model",
                    "frugivory_model",
                    "terrestriality_model"),
         weight = weight %>% round(digits = 2)) %>% 
  select(model, weight) 


# Combine weights and scores
model_comparison_exp <- comp_exp %>%
  select(model,waic, se_waic)%>%
  left_join(weights_exp)%>%
  arrange(desc(weight))

#write_csv(model_comparison_exp, "../saves/model_comparison_exp.csv")

model_comparison_exp <- read_csv("../saves/model_comparison_exp.csv")

model_comparison_exp
```
# Visualize winning models and baseline

## Posterior distributions for predictors
```{r}
model_draws <-  bind_rows(
  dietary_breadth_model %>% posterior_samples(pars = c("b_delay","b_dietary_breadth","b_delay:dietary_breadth"))%>%
    rename(delay = b_delay,
           predictor = b_dietary_breadth,
           interaction = `b_delay:dietary_breadth`)%>%
    mutate(model= "dietary_breadth"),
  vocal_repertoire_model%>% posterior_samples(pars = c("b_delay","b_vocal_repertoire","b_delay:vocal_repertoire"))%>%
    rename(delay = b_delay,
           predictor = b_vocal_repertoire,
           interaction = `b_delay:vocal_repertoire`)%>%
    mutate(model= "vocal_repertoire"),
  base_model%>% posterior_samples(pars = c("b_delay"))%>%
    rename(delay = b_delay)%>%
    mutate(model= "baseline")
)%>%
  pivot_longer(names_to = "estimate", values_to = "value", cols = c(delay,predictor,interaction ))

model_draws %>%
  ggplot(.,aes(x = value, y= model, fill = factor(stat(quantile))))+
  geom_vline(xintercept = 0, lty = 2)+
  stat_density_ridges(
    geom = "density_ridges_gradient",
    calc_ecdf = TRUE,
    quantiles = c(0.025, 0.975)
  ) +
  facet_grid(~estimate)+
  theme_minimal() +
  scale_fill_manual(
   values = c("firebrick", "NA", "firebrick")
  )+
  guides(fill = F)
```
##Correlation between predictor and stm performance

Performance estimate is the species intercept from the baseline model. Not sure if that makes sense.

```{r}
base_model%>% spread_draws(r_species_english[species_english,term])%>%
  filter(term == "Intercept")%>%
  rename(stm_estimate = "r_species_english")%>%
  group_by(species_english)%>%
  median_qi(stm_estimate)%>%
  left_join(species_data%>%
              select(species_english,superordinate_group,vocal_repertoire,dietary_breadth))%>%
  pivot_longer(cols = c("vocal_repertoire","dietary_breadth"), names_to = "variable", values_to = "predictor")%>%
  ggplot(aes(x = predictor, y = stm_estimate, col = superordinate_group))+
  facet_grid(~variable)+
  geom_point(alpha = .5)+
  geom_smooth(method = lm, se = F, lty = 2, col = "grey")+
  stat_cor(method = "pearson", inherit.aes = F, aes(x = predictor, y = stm_estimate), r.accuracy = 0.01, cor.coef.name = "r")+
  theme_minimal()
  
```
