==> analysis/03b_phylogenetic_model_comparison.Rmd <==
---
title: "ManyPrimates1_Phylogenetic_Model_Comparison"
output:
  html_notebook:
    code_folding: hide
    css: style.css
    theme: paper
    toc: yes
    toc_float: yes
---

```{r setup, include=FALSE}
library(ape)
library(tidyverse)
library(readxl)
library(googledrive)
library(brms)
library(broom)
library(ggthemes)
library(tidybayes)
library(ggridges)
library(ggpubr)
library(GGally)
```


# NOTE
Running the models takes a long time (5h on a 64 core HPC) - outputs are too large to put on GitHub (~8GB). If you want to access the model outputs directly, you have to download them via the following link. Code that follows expects files to be in `saves/`.

```{r}
# download zip file from Max Planck cloud storage
options(timeout=3600) # 1h, maybe you need more depending on your internet speed
download.file("https://keeper.mpdl.mpg.de/f/29cc8af41c3a4dd4b8da/?dl=1", 
              destfile = "../saves/saves.zip")
# unpack zip file to get .rds files
unzip("../saves/saves.zip", exdir = "../saves/")
```

If the code doesn't work, you can download the files here: `https://keeper.mpdl.mpg.de/f/29cc8af41c3a4dd4b8da/?dl=1` but you will have to put them manually into the `saves` directory.

# Data processing

## Read in STM data file

```{r}
data <- read_csv("../data/merged_data/ManyPrimates_mp1_merged_data.csv")
```

## Read in species level predictors

```{r}
species_data <- read_xlsx("../data/species_predictors.xlsx", sheet = 1)%>%
  mutate_all(., .funs = tolower)%>%
  select(-contains("Source"), -contains("Comments"))%>%# remove columns with sources and comments
  rename(vocal_repertoire = `vocal_repertoire (# vocalization types)`,
         body_size = `body_size (kg)`)%>%
  mutate_if(is.character, tolower)%>%
  filter(species_english != "allens_swamp_monkey",
         species_english != "hamlyns_monkey") %>%
  mutate(vocal_repertoire = as.numeric(vocal_repertoire),
         group_size = as.numeric(group_size),
         home_range = as.numeric(home_range),
         feeding_budget = as.numeric(feeding_budget),
         day_journey_length= as.numeric(day_journey_length),
         resting_time_percent_activitybudget= as.numeric(resting_time_percent_activitybudget),
         feeding_budget= as.numeric(feeding_budget),
         dietary_breadth= as.numeric(dietary_breadth),
         percent_frugivory= as.numeric(percent_frugivory),
         body_size= as.numeric(body_size),
         home_range_body_site_ratio = home_range/body_size)
```
### Correlation matrix for numeric predictors

```{r, message=F}
species_data%>%
  select(-color_vision,-diet_diversity, -terrestriality, -body_size)%>%
  ggpairs(columns = 5:13,
        diag = list(continous = "density"),
        lower = list(continuous = wrap("smooth", method ="lm", alpha = .3)))+
  scale_color_viridis_d()+
  theme_few()

```

```{r}
ggsave("../figures/phylo_correlations_predictors.png", width = 10, height = 10, scale = 1.5)
```

# Model comparison

## Read in models

```{r}
 base_model  <- readRDS("../saves/phylo_base_model.rds")  
 diurnal_resting_model <- readRDS("../saves/phylo_diurnal_resting_model.rds")
 color_vision_model <- readRDS("../saves/phylo_color_vision_model.rds")
 home_range_feeding_budget_diet_breath_model <- readRDS("../saves/phylo_home_range_feeding_budget_diet_breath_model.rds")
 vocal_repertoire_model <- readRDS("../saves/phylo_vocal_repertoire_model.rds") 
 dietary_breadth_model <- readRDS("../saves/phylo_dietary_breadth_model.rds")
 group_size_model <- readRDS("../saves/phylo_group_size_model.rds")
 day_journey_group_size_model <- readRDS("../saves/phylo_day_journey_group_size_model.rds")
 frugivory_terrestriality_model <- readRDS("../saves/phylo_frugivory_terrestriality_model.rds")
 diet_diversity_model <- readRDS("../saves/phylo_diet_diversity_model.rds")
 home_range_model <- readRDS("../saves/phylo_home_range_model.rds")
 feeding_budget_model <- readRDS("../saves/phylo_feeding_budget_model.rds")
 day_journey_model <- readRDS("../saves/phylo_day_journey_model.rds")
 frugivory_model <- readRDS("../saves/phylo_frugivory_model.rds")
 terrestriality_model <- readRDS("../saves/phylo_terrestriality_model.rds")

```

## Models submitted to modelling challenge

```{r}
# Compute WAIC scores and SE
comp_waic <-  loo_compare(
  base_model,
  diurnal_resting_model,
  color_vision_model,
  home_range_feeding_budget_diet_breath_model,
  vocal_repertoire_model,
  dietary_breadth_model,
  group_size_model,
  day_journey_group_size_model,
  frugivory_terrestriality_model,
  diet_diversity_model,
  criterion = "waic")%>%
  as_tibble(rownames = "model")

# Compute LOO scores and SE
comp_loo <-  loo_compare(
  base_model,
  diurnal_resting_model,
  color_vision_model,
  home_range_feeding_budget_diet_breath_model,
  vocal_repertoire_model,
  dietary_breadth_model,
  group_size_model,
  day_journey_group_size_model,
  frugivory_terrestriality_model,
  diet_diversity_model,
  criterion = "loo")%>%
  as_tibble(rownames = "model")


# Compute WAIC weights
weights_waic <- model_weights(
  base_model,
  diurnal_resting_model,
  color_vision_model,
  home_range_feeding_budget_diet_breath_model,
  vocal_repertoire_model,
  dietary_breadth_model,
  group_size_model,
  day_journey_group_size_model,
  frugivory_terrestriality_model,
  diet_diversity_model, weights = "waic")%>%
  as_tibble() %>%
  rename(waic_weight = value) %>%
  mutate(model  = c("base_model",
                    "diurnal_resting_model",
                    "color_vision_model",
                    "home_range_feeding_budget_diet_breath_model",
                    "vocal_repertoire_model",
                    "dietary_breadth_model",
                    "group_size_model",
                    "day_journey_group_size_model",
                    "frugivory_terrestriality_model",
                    "diet_diversity_model"),
         waic_weight = waic_weight %>% round(digits = 2)) %>%
  select(model, waic_weight)

# Compute LOO weights
weights_loo <- model_weights(
  base_model,
  diurnal_resting_model,
  color_vision_model,
  home_range_feeding_budget_diet_breath_model,
  vocal_repertoire_model,
  dietary_breadth_model,
  group_size_model,
  day_journey_group_size_model,
  frugivory_terrestriality_model,
  diet_diversity_model, weights = "loo")%>%
  as_tibble() %>%
  rename(loo_weight = value) %>%
  mutate(model  = c("base_model",
                    "diurnal_resting_model",
                    "color_vision_model",
                    "home_range_feeding_budget_diet_breath_model",
                    "vocal_repertoire_model",
                    "dietary_breadth_model",
                    "group_size_model",
                    "day_journey_group_size_model",
                    "frugivory_terrestriality_model",
                    "diet_diversity_model"),
         loo_weight = loo_weight %>% round(digits = 2)) %>%
  select(model, loo_weight)


# Combine weights and scores
model_comparison <- comp_waic %>% select(model,waic, se_waic)%>%
  left_join(weights_waic)%>%
  left_join(comp_loo%>% select(model,looic, se_looic))%>%
  left_join(weights_loo)%>%
  arrange(desc(waic_weight))

write_csv(model_comparison, "../saves/phylogenetic_model_comparison.csv")

model_comparison <- read_csv("../saves/phylogenetic_model_comparison.csv")

model_comparison
```
## Exploratory analysis: Include main effects models for all predictors

Some models were submitted with interaction terms between variables. The additional models below were built so that each predictor that was submitted as part of an interaction is also included in a model with only that predictor. 

```{r}
# Compute WAIC scores and SE
comp_exp_waic <-  loo_compare(
  base_model,    # models from modelling challnege
  diurnal_resting_model,
  color_vision_model,
  home_range_feeding_budget_diet_breath_model,
  vocal_repertoire_model,
  dietary_breadth_model,
  group_size_model,
  day_journey_group_size_model,
  frugivory_terrestriality_model,
  diet_diversity_model,

  home_range_model,      # main effects models
  feeding_budget_model,
  day_journey_model,
  frugivory_model,
  terrestriality_model,

  criterion = "waic")%>%
  as_tibble(rownames = "model")

# Compute LOO scores and SE
comp_exp_loo <-  loo_compare(
  base_model,    # models from modelling challnege
  diurnal_resting_model,
  color_vision_model,
  home_range_feeding_budget_diet_breath_model,
  vocal_repertoire_model,
  dietary_breadth_model,
  group_size_model,
  day_journey_group_size_model,
  frugivory_terrestriality_model,
  diet_diversity_model,

  home_range_model,      # main effects models
  feeding_budget_model,
  day_journey_model,
  frugivory_model,
  terrestriality_model,

  criterion = "loo")%>%
  as_tibble(rownames = "model")

# Compute WAIC weights
weights_exp_waic <- model_weights(
  base_model,
  diurnal_resting_model,
  color_vision_model,
  home_range_feeding_budget_diet_breath_model,
  vocal_repertoire_model,
  dietary_breadth_model,
  group_size_model,
  day_journey_group_size_model,
  frugivory_terrestriality_model,
  diet_diversity_model,
  home_range_model,
  feeding_budget_model,
  day_journey_model,
  frugivory_model,
  terrestriality_model,
  weights = "waic")%>%
  as_tibble() %>%
  rename(waic_weight = value) %>%
  mutate(model  = c("base_model",
                    "diurnal_resting_model",
                    "color_vision_model",
                    "home_range_feeding_budget_diet_breath_model",
                    "vocal_repertoire_model",
                    "dietary_breadth_model",
                    "group_size_model",
                    "day_journey_group_size_model",
                    "frugivory_terrestriality_model",
                    "diet_diversity_model",
                    "home_range_model",
                    "feeding_budget_model",
                    "day_journey_model",
                    "frugivory_model",
                    "terrestriality_model"),
         waic_weight = waic_weight %>% round(digits = 2)) %>%
  select(model, waic_weight)

# Compute LOO weights
weights_exp_loo <- model_weights(
  base_model,
  diurnal_resting_model,
  color_vision_model,
  home_range_feeding_budget_diet_breath_model,
  vocal_repertoire_model,
  dietary_breadth_model,
  group_size_model,
  day_journey_group_size_model,
  frugivory_terrestriality_model,
  diet_diversity_model,
  home_range_model,
  feeding_budget_model,
  day_journey_model,
  frugivory_model,
  terrestriality_model,
  weights = "loo")%>%
  as_tibble() %>%
  rename(loo_weight = value) %>%
  mutate(model  = c("base_model",
                    "diurnal_resting_model",
                    "color_vision_model",
                    "home_range_feeding_budget_diet_breath_model",
                    "vocal_repertoire_model",
                    "dietary_breadth_model",
                    "group_size_model",
                    "day_journey_group_size_model",
                    "frugivory_terrestriality_model",
                    "diet_diversity_model",
                    "home_range_model",
                    "feeding_budget_model",
                    "day_journey_model",
                    "frugivory_model",
                    "terrestriality_model"),
         loo_weight = loo_weight %>% round(digits = 2)) %>%
  select(model, loo_weight)



# Combine weights and scores
model_comparison_exp <- comp_exp_waic %>% select(model,waic, se_waic)%>%
  left_join(weights_exp_waic)%>%
  left_join(comp_exp_loo%>% select(model,looic, se_looic))%>%
  left_join(weights_exp_loo)%>%
  arrange(desc(waic_weight))

write_csv(model_comparison_exp, "../saves/phylogenetic_model_comparison_exploratory.csv")

model_comparison_exp <- read_csv("../saves/phylogenetic_model_comparison_exploratory.csv")

model_comparison_exp

```

## Exploratory analysis: importance of phylogenetic tree in random effects

Here we compare the baseline phylogenetic model to the same model without the phylogenetic tree in the random effect structure. 

### Read in non-phylogenetic model

To access the model output, download it here: `https://keeper.mpdl.mpg.de/d/d258f3c7a62948b58341/`

```{r}
no_phylogeny_model <- readRDS("../saves/phylo_no_phylogeny_model.rds")
```

### Model comparison 

```{r}
phylo_comp_waic <-  loo_compare(
  base_model,
  no_phylogeny_model,
  criterion = c("waic"))%>%
  as_tibble(rownames = "model")

phylo_comp_loo <-  loo_compare(
  base_model,
  no_phylogeny_model,
  criterion = c("loo"))%>%
  as_tibble(rownames = "model")

phylo_weights_waic <- model_weights(
  base_model,
  no_phylogeny_model,
  weights = "waic")%>%
  as_tibble() %>%
  rename(waic_weight = value) %>%
  mutate(model  = c("base_model", "no_phylogeny_model"),
         waic_weight = waic_weight %>% round(digits = 4)) %>%
  select(model, waic_weight)

phylo_weights_loo <- model_weights(
  base_model,
  no_phylogeny_model,
  weights = "loo")%>%
  as_tibble() %>%
  rename(loo_weight = value) %>%
  mutate(model  = c("base_model", "no_phylogeny_model"),
         loo_weight = loo_weight %>% round(digits = 4)) %>%
  select(model, loo_weight)

# Combine weights and scores
phylo_model_comparison <- phylo_comp_waic %>%select(model,waic, se_waic)%>%
  left_join(phylo_weights_waic)%>%
  left_join(phylo_comp_loo %>%select(model,looic, se_looic))%>%
  left_join(phylo_weights_loo)%>%
  arrange(desc(waic_weight))

write_csv(phylo_model_comparison, "../saves/phylogenetic_model_comparison_phylogeny.csv")

phylo_model_comparison <- read_csv("../saves/phylogenetic_model_comparison_phylogeny.csv")

```

# Visualize predictor models and baseline

## Posterior distributions for predictors
```{r}
model_draws <-  bind_rows(
  vocal_repertoire_model%>% posterior_samples(pars = c("b_delay1","b_vocal_repertoire","b_delay1:vocal_repertoire"))%>%
    rename(delay = b_delay1,
           predictor = b_vocal_repertoire,
           `delay x predictor` = `b_delay1:vocal_repertoire`)%>%
    mutate(model= "vocal_repertoire"),
  dietary_breadth_model%>% posterior_samples(pars = c("b_delay1","b_dietary_breadth","b_delay1:dietary_breadth"))%>%
    rename(delay = b_delay1,
           predictor = b_dietary_breadth,
           `delay x predictor` = `b_delay1:dietary_breadth`)%>%
    mutate(model= "dietary_breadth"),
  base_model%>% posterior_samples(pars = c("b_delay1"))%>%
    rename(delay = b_delay1)%>%
    mutate(model= "baseline")
)%>%
  pivot_longer(names_to = "estimate", values_to = "value", cols = c(delay,predictor,`delay x predictor` ))

model_draws %>%
  ggplot(.,aes(x = value, y= model, fill = factor(stat(quantile))))+
  geom_vline(xintercept = 0, lty = 2)+
  stat_density_ridges(
    geom = "density_ridges_gradient",
    calc_ecdf = TRUE,
    quantiles = c(0.025, 0.975)
  ) +
  facet_grid(~estimate)+
  theme_minimal() +
  scale_fill_manual(
   values = c("firebrick", "NA", "firebrick")
  )+
  guides(fill = F)
```

## Model predictions for delay

```{r}

model_pred <-  bind_rows(
  vocal_repertoire_model%>% 
  posterior_samples(pars = c("b_Intercept","b_delay1","b_vocal_repertoire","b_delay1:vocal_repertoire"), subset = runif(1000, 0, 20000))%>%
  rownames_to_column(var = "iteration")%>%
  rename(intercept = b_Intercept, 
         b_delay = b_delay1,
         main_effect = b_vocal_repertoire,
         interaction = `b_delay1:vocal_repertoire`)%>%
  expand_grid(predictor = seq(-1,1, length=100))%>%
  mutate(no_delay = intercept + b_delay * 1 + main_effect * predictor + interaction * predictor * -1,
         delay = intercept + b_delay * 0 + main_effect * predictor + interaction * predictor * 0)%>%
  select(-b_delay,-intercept, - interaction, -main_effect)%>%
  pivot_longer(cols = c(no_delay, delay), names_to = "delay", values_to = "prediction")%>%
  mutate(prediction = plogis(prediction),
         delay = factor(delay, levels = c("no_delay", "delay")),
         model= "vocal_repertoire"),
  
  dietary_breadth_model%>% 
  posterior_samples(pars = c("b_Intercept","b_delay1","b_dietary_breadth","b_delay1:dietary_breadth"), subset = runif(1000, 0, 20000))%>%
  rownames_to_column(var = "iteration")%>%
  rename(intercept = b_Intercept, 
         b_delay = b_delay1,
         main_effect = b_dietary_breadth,
         interaction = `b_delay1:dietary_breadth`)%>%
  expand_grid(predictor = seq(-1,1, length=100))%>%
  mutate(no_delay = intercept + b_delay * 1 + main_effect * predictor + interaction * predictor * -1,
         delay = intercept + b_delay * 0 + main_effect * predictor + interaction * predictor * 0)%>%
  select(-b_delay,-intercept, - interaction, -main_effect)%>%
  pivot_longer(cols = c(no_delay, delay), names_to = "delay", values_to = "prediction")%>%
  mutate(prediction = plogis(prediction),
         delay = factor(delay, levels = c("no_delay", "delay")),
         model= "dietary_breadth")
  )

model_pred_map <- bind_rows(
  
  fixef(vocal_repertoire_model)%>%
  as_tibble(rownames = "var")%>%
  select(var, Estimate)%>%
  pivot_wider(names_from = var, values_from = Estimate)%>%
  rename(intercept = Intercept, 
         b_delay = delay1,
         main_effect = vocal_repertoire,
         interaction = `delay1:vocal_repertoire`)%>%
  expand_grid(predictor = seq(-1,1, length=100))%>%
  mutate(no_delay = intercept + b_delay * 1 + main_effect * predictor + interaction * predictor * -1,
         delay = intercept + b_delay * 0 + main_effect * predictor + interaction * predictor * 0)%>%
  select(-b_delay,-intercept, - interaction, -main_effect)%>%
  pivot_longer(cols = c(no_delay, delay), names_to = "delay", values_to = "prediction")%>%
  mutate(prediction = plogis(prediction),
         delay = factor(delay, levels = c("no_delay", "delay")),
         model= "vocal_repertoire"),
 
  fixef(dietary_breadth_model)%>%
  as_tibble(rownames = "var")%>%
  select(var, Estimate)%>%
  pivot_wider(names_from = var, values_from = Estimate)%>%
  rename(intercept = Intercept, 
         b_delay = delay1,
         main_effect = dietary_breadth,
         interaction = `delay1:dietary_breadth`)%>%
  expand_grid(predictor = seq(-1,1, length=100))%>%
  mutate(no_delay = intercept + b_delay * 1 + main_effect * predictor + interaction * predictor * -1,
         delay = intercept + b_delay * 0 + main_effect * predictor + interaction * predictor * 0)%>%
  select(-b_delay,-intercept, - interaction, -main_effect)%>%
  pivot_longer(cols = c(no_delay, delay), names_to = "delay", values_to = "prediction")%>%
  mutate(prediction = plogis(prediction),
         delay = factor(delay, levels = c("no_delay", "delay")),
         model= "dietary_breadth"),
   
)
  


ggplot(model_pred,aes(x = predictor, y= prediction, col = delay))+
  geom_hline(yintercept = 1/3, lty = 2, alpha = .5)+
  geom_line(alpha = 0.025, aes(group = interaction(delay,iteration)))+
  geom_line(data = model_pred_map, aes(x = predictor, y= prediction, col = delay),size = 1.5)+
  facet_grid(model~delay)+
  ylim(0,1)+
  theme_minimal() +
  guides(col = F)+
  scale_color_colorblind()
```
