==> analysis/03c_phylogenetic_signal.Rmd <==
---
title: "ManyPrimates1_Phylogenetic_Signal"
output:
  html_notebook:
    code_folding: hide
    css: style.css
    theme: paper
    toc: yes
    toc_float: yes
---

```{r setup, include=FALSE}
library(ape)
library(brms)
library(parallel)
library(phytools)
library(tidyverse)
library(readxl)
library(ggdist)

# Load species data
species_data <- read_xlsx("../data/species_predictors.xlsx", sheet = 1)%>%
  mutate_all(., .funs = tolower)%>%
  select(-contains("Source"), -contains("Comments"))%>%# remove columns with sources and comments
  rename(vocal_repertoire = `vocal_repertoire (# vocalization types)`,
         body_size = `body_size (kg)`)%>%
  mutate_if(is.character, tolower)%>%
  filter(species_english != "allens_swamp_monkey",
         species_english != "hamlyns_monkey") %>%
  mutate(vocal_repertoire = as.numeric(vocal_repertoire),
         group_size = as.numeric(group_size),
         home_range = as.numeric(home_range),
         feeding_budget = as.numeric(feeding_budget),
         day_journey_length= as.numeric(day_journey_length),
         resting_time_percent_activitybudget= as.numeric(resting_time_percent_activitybudget),
         feeding_budget= as.numeric(feeding_budget),
         dietary_breadth= as.numeric(dietary_breadth),
         percent_frugivory= as.numeric(percent_frugivory),
         body_size= as.numeric(body_size),
         home_range_body_site_ratio = home_range/body_size)%>%
  mutate(resting_time_percent_activitybudget = scale(resting_time_percent_activitybudget, center = T, scale = T),
         home_range_body_site_ratio = scale(home_range_body_site_ratio, center = T, scale = T),
         home_range = scale(home_range, center = T, scale = T),
         feeding_budget = scale(feeding_budget, center = T, scale = T),
         dietary_breadth = scale(dietary_breadth, center = T, scale = T),
         vocal_repertoire = scale(vocal_repertoire, center = T, scale = T),
         group_size = scale(group_size, center = T, scale = T),
         day_journey_length = scale(day_journey_length, center = T, scale = T),
         percent_frugivory = scale(percent_frugivory, center = T, scale = T),
         feeding_budget = scale(feeding_budget, center = T, scale =T))

# Load phylo data
data <- read_csv("../data/merged_data/ManyPrimates_mp1_merged_data.csv")
phylo_data <- data %>%
  rename(species_english = species)%>% # create a common column to merge by
  filter(species_english != "allens_swamp_monkey",
         species_english != "hamlyns_monkey")

# Load tree
tree <- read.tree("../phylo/mp_species.tree")

# Species sanity checking
data_species <- phylo_data$species_english %>% unique
species_species <- species_data$species_english %>% unique
tree_species <- tree$tip.label
stopifnot(length(setdiff(species_species, data_species))== 0)
stopifnot(length(setdiff(data_species, species_species))== 0)
stopifnot(length(setdiff(species_species, tree_species)) == 0)
stopifnot(length(setdiff(data_species, tree_species)) == 0)
tree <- keep.tip(tree, data_species)

# Compute lambda and kappa values for the various predictor variables
predictor_vars <- colnames(species_data)[5:17]
signal <- data_frame(variable=c(), lambda=c(), kappa=c())
for(pred in predictor_vars) {
  if(typeof(species_data[[pred]]) != "double") {
    print(c("Skipping", pred))
    next
  }
  predictors <- setNames(as.vector(species_data[[pred]]), species_data$species_english)
  # Calls to `phylosig` below complain "x has no names; assuming x is in the same order as tree$tip.label"
  # This is clearly false (see `setNames` above).
  # To reassure myself that this is some spurious error on `phytools`' part...
  stopifnot(length(intersect(names(predictors), tree$tip.label)) == length(tree$tip.label))
  L <- phylosig(tree, predictors, method="lambda")$lambda
  K <- phylosig(tree, predictors, method="K")[1]
  signal <- add_row(signal, variable=pred, lambda=L, kappa=K)
}

# Compute lambda and kappa values for base model predictions
phylo_base_model <- readRDS("../saves/phylo_base_model.rds")

do_prediction <- function(delay) {
  new_data <- data_frame(species_english=species_data$species_english,
                         site=1:nrow(species_data), subject_site=1:nrow(species_data),
                         delay=delay, trial=0)
  predictors <- posterior_linpred(phylo_base_model, newdata=new_data,
                                  allow_new_levels=TRUE,
                                  sample_new_levels="old_levels")
  predictors <- setNames(as.vector(colMeans(predictors)), new_data$species_english)
  return(predictors)
}

options(warn=2)
results <- mclapply(rep(c(-1, 0, 1), 3333), do_prediction, mc.cores=detectCores())
options(warn=0)

for(i in 1:length(results)) {
  print(results[[i]])
  L <- phylosig(tree, results[[i]], method="lambda")$lambda
  K <- phylosig(tree, results[[i]], method="K")[1]
  signal <- add_row(signal, variable="model", lambda=L, kappa=K)
}

write_csv(signal, "../saves/phylogenetic_signal.csv")

signal <- read_csv("../saves/phylogenetic_signal.csv")

# Mean and HDI for phylogenetic signals
signal %>%
  filter(variable == "model")%>%
  pivot_longer(cols = -variable, names_to = "index", values_to = "value")%>%
  group_by(index)%>%
  mean_hdci(value)

```

# Visualisation of indices 

```{r}
signal%>%
  filter(variable == "model")%>%
  pivot_longer(cols = -variable, names_to = "index", values_to = "value")%>%
  ggplot(aes(x = value))+
  geom_density()+
  facet_grid(~index)+
  theme_minimal()+
  xlim(0,1)
```
