==> analysis/04_confirmatory_analysis.Rmd <==
---
title: "ManyPrimates1_Confirmatory_Analysis"
output:
  html_notebook:
    code_folding: hide
    css: style.css
    theme: paper
    toc: yes
    toc_float: yes
---

```{r setup, include=FALSE}
library(ape)
library(tidyverse)
library(readxl)
library(googledrive)
library(brms)
library(broom)
library(ggthemes)
library(tidybayes)
library(ggridges)
library(ggpubr)
library(BayesFactor)
```


# NOTE
Running the models takes quite some time (2h on a HPC) - outputs are too large to put on GitHub (~2GB). If you want to access the model outputs directly, you have to download them via the following link. Code that follows expects files to be in `saves/`.

```{r}
# download zip file from Max Planck cloud storage
options(timeout=3600) # 1h, maybe you need more depending on your internet speed
download.file("https://keeper.mpdl.mpg.de/f/9379c00231fc42e38920/?dl=1", 
              destfile = "../saves/saves.zip")
# unpack zip file to get .rds files
unzip("../saves/saves.zip", exdir = "../saves/")
```

If the code doesn't work, you can download the files here: `https://keeper.mpdl.mpg.de/f/9379c00231fc42e38920/?dl=1` but you will have to put them manually into the `saves` directory.

# Data processing

## Read in STM data file

Read in data, re-code condition and scale predictors. 

```{r}
data <- read_csv("../data/merged_data/ManyPrimates_mp1_merged_data.csv")%>%
  mutate(delay = if_else(condition == "short", -1,
                         if_else(condition == "medium", 0, 1)),
         trial = scale(trial, center = T, scale = T),
         norm_age = scale(norm_age, center = T, scale = T),
         cup_distance = scale(cup_distance, center = T, scale = T),
         board_size = scale(board_size, center = T, scale = T))
```

# Confirmatory analysis

## Comparison to chance

```{r}
data %>%
  group_by(condition, subject_site) %>%
  summarise(correct = mean(correct)) %>%
  summarise(correct = list(correct)) %>%
  group_by(condition)%>%
  mutate(mean= mean(unlist(correct)),
         BF = round(extractBF(ttestBF(unlist(correct), mu = 1/3))$bf,2))%>%
  select(-correct)
```

## Full model, numeric delay

```{r}
numeric_delay_model_filename <- "../saves/conf_numeric_delay_model.rds"
if(file.exists(numeric_delay_model_filename)) {
    numeric_delay_model <- readRDS(numeric_delay_model_filename)
} else {
    numeric_delay_model <- brm(correct ~ delay +
                      task_experience + norm_age + cup_distance + board_size + trial +
                      (1 + delay + trial | subject_site + site + species),
                      data=data,
                      family=bernoulli(),
                      chains = 4,
                      iter= 5000,
                      cores= 4,
                      prior = prior(skew_normal(0, 1.5, 3), class=Intercept) +
                              prior(skew_normal(-1.4, 2, -3), class=b, coef=delay) +
                              prior(normal(0, 1), class=b, coef=board_size) +
                              prior(normal(0, 1), class=b, coef=trial) +
                              prior(normal(0, 1), class=b, coef=cup_distance) +
                              prior(normal(0, 1), class=b, coef=norm_age) +
                              prior(normal(0, 2), class=b, coef=task_experienceyes) +
                              prior(normal(0, 1), class=sd, coef=Intercept, group=subject_site) +
                              prior(normal(0, 1), class=sd, coef=Intercept, group=site) +
                              prior(normal(0, 1), class=sd, coef=Intercept, group=species) +
                              prior(normal(0, 0.5), class=sd, coef=delay, group=subject_site) +
                              prior(normal(0, 0.5), class=sd, coef=delay, group=site) +
                              prior(normal(0, 0.5), class=sd, coef=delay, group=species) +
                              prior(normal(0, 0.5), class=sd, coef=trial, group=subject_site) +
                              prior(normal(0, 0.5), class=sd, coef=trial, group=site) +
                              prior(normal(0, 0.5), class=sd, coef=trial, group=species) +
                              prior("lkj(1)", class = "cor")
                   )
    summary(numeric_delay_model)
    numeric_delay_model <- add_criterion(numeric_delay_model, c("loo", "waic"))
    saveRDS(numeric_delay_model, numeric_delay_model_filename)
}
```

## Full model, three-level factor delay

```{r}
factor_delay_model_filename <- "../saves/conf_factor_delay_model.rds"
if(file.exists(factor_delay_model_filename)) {
    factor_delay_model <- readRDS(factor_delay_model_filename)
} else {
    data$factor_delay <- factor(data$delay, c(0, -1, 1))
    factor_delay_model <- brm(correct ~ factor_delay +
                      task_experience + norm_age + cup_distance + board_size + trial +
                      (1 + factor_delay + trial | subject_site + site + species),
                      data=data,
                      family=bernoulli(),
                      chains = 4,
                      iter= 5000,
                      cores= 4,
                      prior = prior(skew_normal(0, 1.5, 3), class=Intercept) +
                              prior(skew_normal(1.4, 2, 3), class=b, coef=factor_delayM1) +
                              prior(skew_normal(-1.4, 2, -3), class=b, coef=factor_delay1) +
                              prior(normal(0, 1), class=b, coef=board_size) +
                              prior(normal(0, 1), class=b, coef=trial) +
                              prior(normal(0, 1), class=b, coef=cup_distance) +
                              prior(normal(0, 1), class=b, coef=norm_age) +
                              prior(normal(0, 2), class=b, coef=task_experienceyes) +
                              prior(normal(0, 1), class=sd, coef=Intercept, group=subject_site) +
                              prior(normal(0, 1), class=sd, coef=Intercept, group=site) +
                              prior(normal(0, 1), class=sd, coef=Intercept, group=species) +
                              prior(normal(0, 0.5), class=sd, coef=factor_delayM1, group=subject_site) +
                              prior(normal(0, 0.5), class=sd, coef=factor_delayM1, group=site) +
                              prior(normal(0, 0.5), class=sd, coef=factor_delayM1, group=species) +
                              prior(normal(0, 0.5), class=sd, coef=factor_delay1, group=subject_site) +
                              prior(normal(0, 0.5), class=sd, coef=factor_delay1, group=site) +
                              prior(normal(0, 0.5), class=sd, coef=factor_delay1, group=species) +
                              prior(normal(0, 0.5), class=sd, coef=trial, group=subject_site) +
                              prior(normal(0, 0.5), class=sd, coef=trial, group=site) +
                              prior(normal(0, 0.5), class=sd, coef=trial, group=species) +
                              prior("lkj(1)", class = "cor")
                   )
    summary(factor_delay_model)
    factor_delay_model <- add_criterion(factor_delay_model, c("loo", "waic"))
    saveRDS(factor_delay_model, factor_delay_model_filename)
}
```

## Full model, two-level factor delay

```{r}
binary_delay_model_filename <- "../saves/conf_binary_delay_model.rds"
if(file.exists(binary_delay_model_filename)) {
    binary_delay_model <- readRDS(binary_delay_model_filename)
} else {
    data$binary_delay <- as.factor(as.integer(data$delay > -1))
    binary_delay_model <- brm(correct ~ binary_delay +
                      task_experience + norm_age + cup_distance + board_size + trial +
                      (1 + binary_delay + trial | subject_site + site + species),
                      data=data,
                      family=bernoulli(),
                      chains = 4,
                      iter= 5000,
                      cores= 4,
                      prior = prior(skew_normal(0, 1.5, 3), class=Intercept) +
                              prior(skew_normal(-1.4, 2, -3), class=b, coef=binary_delay1) +
                              prior(normal(0, 1), class=b, coef=board_size) +
                              prior(normal(0, 1), class=b, coef=trial) +
                              prior(normal(0, 1), class=b, coef=cup_distance) +
                              prior(normal(0, 1), class=b, coef=norm_age) +
                              prior(normal(0, 2), class=b, coef=task_experienceyes) +
                              prior(normal(0, 1), class=sd, coef=Intercept, group=subject_site) +
                              prior(normal(0, 1), class=sd, coef=Intercept, group=site) +
                              prior(normal(0, 1), class=sd, coef=Intercept, group=species) +
                              prior(normal(0, 0.5), class=sd, coef=binary_delay1, group=subject_site) +
                              prior(normal(0, 0.5), class=sd, coef=binary_delay1, group=site) +
                              prior(normal(0, 0.5), class=sd, coef=binary_delay1, group=species) +
                              prior(normal(0, 0.5), class=sd, coef=trial, group=subject_site) +
                              prior(normal(0, 0.5), class=sd, coef=trial, group=site) +
                              prior(normal(0, 0.5), class=sd, coef=trial, group=species) +
                              prior("lkj(1)", class = "cor")
                   )
    summary(binary_delay_model)
    binary_delay_model <- add_criterion(binary_delay_model, c("loo", "waic"))
    saveRDS(binary_delay_model, binary_delay_model_filename)
}
```

## Null model

```{r}
null_model_filename <- "../saves/conf_null_model.rds"
if(file.exists(null_model_filename)) {
    null_model <- readRDS(null_model_filename)
} else {
    null_model <- brm(correct ~ task_experience + norm_age + cup_distance + board_size + trial +
                      (1 + trial | subject_site + site + species),
                      data=data,
                      family=bernoulli(),
                      chains = 4,
                      iter= 5000,
                      cores= 4,
                      prior = prior(skew_normal(0, 1.5, 3), class=Intercept) +
                              prior(normal(0, 1), class=b, coef=board_size) +
                              prior(normal(0, 1), class=b, coef=trial) +
                              prior(normal(0, 1), class=b, coef=cup_distance) +
                              prior(normal(0, 1), class=b, coef=norm_age) +
                              prior(normal(0, 2), class=b, coef=task_experienceyes) +
                              prior(normal(0, 1), class=sd, coef=Intercept, group=subject_site) +
                              prior(normal(0, 1), class=sd, coef=Intercept, group=site) +
                              prior(normal(0, 1), class=sd, coef=Intercept, group=species) +
                              prior(normal(0, 0.5), class=sd, coef=trial, group=subject_site) +
                              prior(normal(0, 0.5), class=sd, coef=trial, group=site) +
                              prior(normal(0, 0.5), class=sd, coef=trial, group=species) +
                              prior("lkj(1)", class = "cor")
                 )

    summary(null_model)
    null_model <- add_criterion(null_model, c("loo", "waic"))
    saveRDS(null_model, null_model_filename)
}
```

## Read in model outputs

```{r}
numeric_delay_model <- readRDS("../saves/conf_numeric_delay_model.rds")
factor_delay_model <- readRDS("../saves/conf_factor_delay_model.rds")
binary_delay_model <- readRDS("../saves/conf_binary_delay_model.rds")
null_model <- readRDS("../saves/conf_null_model.rds")
```


# Model comparison

```{r}
model_comparison_filename <- "../saves/confirmatory_model_comparison.csv"
if(file.exists(model_comparison_filename)) {
    model_comparison <- read_csv(model_comparison_filename)
} else {
    comp_waic <-  loo_compare(numeric_delay_model, factor_delay_model, binary_delay_model,
			      null_model, criterion = c("waic")) %>%
                      as_tibble(rownames = "model")

    comp_loo <-  loo_compare(numeric_delay_model, factor_delay_model, binary_delay_model,
			      null_model, criterion = c("loo")) %>%
                      as_tibble(rownames = "model")

    weights_waic <- model_weights(numeric_delay_model, factor_delay_model, binary_delay_model,
				  null_model, weights = "waic") %>%
        as_tibble() %>%
        rename(waic_weight = value) %>%
        mutate(model  = c("numeric_delay_model", "factor_delay_model", "binary_delay_model",
			  "null_model"), waic_weight = waic_weight %>% round(digits = 4)) %>%
        select(model, waic_weight)

    weights_loo <- model_weights(numeric_delay_model, factor_delay_model, binary_delay_model,
				  null_model, weights = "loo") %>%
        as_tibble() %>%
        rename(loo_weight = value) %>%
        mutate(model  = c("numeric_delay_model", "factor_delay_model", "binary_delay_model",
			  "null_model"), loo_weight = loo_weight %>% round(digits = 4)) %>%
        select(model, loo_weight)

    # Combine weights and scores
    model_comparison <- comp_waic %>%select(model,waic, se_waic)%>%
        left_join(weights_waic)%>%
        left_join(comp_loo %>%select(model,looic, se_looic))%>%
        left_join(weights_loo)%>%
        arrange(desc(waic_weight))

    write_csv(model_comparison, "../saves/confirmatory_model_comparison.csv")
}
```

# Visualize full model

## Posterior distributions for fixed effects
```{r}
model_draws <- factor_delay_model%>% posterior_samples(pars = c("b_factor_delayM1","b_factor_delay1", "b_task_experienceyes", "b_norm_age", "b_cup_distance", "b_board_size", "b_trial"))%>%
    mutate(model= "baseline")%>%
  pivot_longer(names_to = "estimate", values_to = "value", cols = c(b_factor_delayM1,b_factor_delay1,b_task_experienceyes,b_norm_age, b_cup_distance, b_board_size, b_trial ))%>%
  mutate(estimate = str_remove(estimate,"b_"),
         estimate = recode(estimate, 
                           factor_delayM1 = "delay_short",
                           factor_delay1 = "delay_long"))

model_draws %>%
  ggplot(.,aes(x = value, y= estimate, fill = factor(stat(quantile))))+
  geom_vline(xintercept = 0, lty = 2)+
  stat_density_ridges(
    geom = "density_ridges_gradient",
    calc_ecdf = TRUE,
    quantiles = c(0.025, 0.975)
  ) +
  theme_minimal() +
  scale_fill_manual(
   values = c("firebrick", "NA", "firebrick")
  )+
  guides(fill = F)
```

```{r}
#ggsave("../figures/confirmatory_analysis_main_effects.png")
```

# Exploratory analysis for age

## Interaction model

In the pilot, we included age in an interaction with delay. We add this model here as part of the exploratory analysis. 

```{r}
int_model_filename <- "../saves/conf_int_model.rds"
if(file.exists(int_model_filename)) {
    int_model <- readRDS(int_model_filename)
} else {
    data$binary_delay <- as.factor(as.integer(data$delay > -1))
    int_model <- brm(correct ~ binary_delay * norm_age +
                     task_experience + cup_distance + board_size + trial +
                     (1 + binary_delay + trial | subject_site + site + species),
                     data=data,
                     family=bernoulli(),
                     chains = 4,
                     iter= 5000,
                     cores= 4,
                     prior = prior(skew_normal(0, 1.5, 3), class=Intercept) +
                             prior(skew_normal(-1.4, 2, -3), class=b, coef=binary_delay1) +
                             prior(normal(0, 1), class=b, coef=board_size) +
                             prior(normal(0, 1), class=b, coef=trial) +
                             prior(normal(0, 1), class=b, coef=cup_distance) +
                             prior(normal(0, 1), class=b, coef=norm_age) +
                             prior(normal(0, 1), class=b, coef=`binary_delay1:norm_age`) +
                             prior(normal(0, 2), class=b, coef=task_experienceyes) +
                             prior(normal(0, 1), class=sd, coef=Intercept, group=subject_site) +
                             prior(normal(0, 1), class=sd, coef=Intercept, group=site) +
                             prior(normal(0, 1), class=sd, coef=Intercept, group=species) +
                             prior(normal(0, 0.5), class=sd, coef=binary_delay1, group=subject_site) +
                             prior(normal(0, 0.5), class=sd, coef=binary_delay1, group=site) +
                             prior(normal(0, 0.5), class=sd, coef=binary_delay1, group=species) +
                             prior(normal(0, 0.5), class=sd, coef=trial, group=subject_site) +
                             prior(normal(0, 0.5), class=sd, coef=trial, group=site) +
                             prior(normal(0, 0.5), class=sd, coef=trial, group=species) +
                             prior("lkj(1)", class = "cor")
                  )

    summary(int_model)

    int_model <- add_criterion(int_model, c("loo", "waic"))
    saveRDS(int_model, int_model_filename)
}

```

## Quadratic model

We also add a model that assumes a non-linear relation between age and performance. 

```{r}
quad_model_filename <- "../saves/conf_quad_model.rds"
if(file.exists(quad_model_filename)) {
    quad_model <- readRDS(quad_model_filename)
} else {
    data$binary_delay <- as.factor(as.integer(data$delay > -1))
    quad_model <- brm(correct ~ binary_delay + norm_age + I(norm_age^2) +
                      task_experience + cup_distance + board_size + trial +
                     (1 + binary_delay + trial | subject_site + site + species),
                     data=data,
                     family=bernoulli(),
                     chains = 4,
                     iter= 5000,
                     cores= 4,
                     prior = prior(skew_normal(0, 1.5, 3), class=Intercept) +
                             prior(skew_normal(-1.4, 2, -3), class=b, coef=binary_delay1) +
                             prior(normal(0, 1), class=b, coef=board_size) +
                             prior(normal(0, 1), class=b, coef=trial) +
                             prior(normal(0, 1), class=b, coef=cup_distance) +
                             prior(normal(0, 1), class=b, coef=norm_age) +
                             prior(normal(0, 1), class=b, coef=Inorm_ageE2) +
                             prior(normal(0, 2), class=b, coef=task_experienceyes) +
                             prior(normal(0, 1), class=sd, coef=Intercept, group=subject_site) +
                             prior(normal(0, 1), class=sd, coef=Intercept, group=site) +
                             prior(normal(0, 1), class=sd, coef=Intercept, group=species) +
                             prior(normal(0, 0.5), class=sd, coef=binary_delay1, group=subject_site) +
                             prior(normal(0, 0.5), class=sd, coef=binary_delay1, group=site) +
                             prior(normal(0, 0.5), class=sd, coef=binary_delay1, group=species) +
                             prior(normal(0, 0.5), class=sd, coef=trial, group=subject_site) +
                             prior(normal(0, 0.5), class=sd, coef=trial, group=site) +
                             prior(normal(0, 0.5), class=sd, coef=trial, group=species) +
                             prior("lkj(1)", class = "cor")
                  )

    summary(quad_model)
    quad_model <- add_criterion(quad_model, c("loo", "waic"))
    saveRDS(quad_model, quad_model_filename)
}
```

## Read in model files

```{r}
int_model <- readRDS("../saves/conf_int_model.rds")
quad_model <- readRDS("../saves/conf_quad_model.rds")

```

## Model comparison

```{r}
model_comparison_exp_filename <- "../saves/confirmatory_model_comparison_exploratory.csv"
if(!file.exists(model_comparison_exp_filename)) {

 comp_waic_exp <-  loo_compare(
   binary_delay_model,
   int_model,
   quad_model,
   criterion = c("waic"))%>%
   as_tibble(rownames = "model")

 comp_loo_exp <-  loo_compare(
   binary_delay_model,
   int_model,
   quad_model,
   criterion = c("loo"))%>%
   as_tibble(rownames = "model")

 weights_waic_exp <- model_weights(binary_delay_model, int_model,quad_model, weights = "waic")%>%
   as_tibble() %>%
   rename(waic_weight = value) %>%
   mutate(model  = c("binary_delay_model", "int_model","quad_model"),
          waic_weight = waic_weight %>% round(digits = 4)) %>%
   select(model, waic_weight)

 weights_loo_exp <- model_weights(binary_delay_model, int_model, quad_model, weights = "loo")%>%
   as_tibble() %>%
   rename(loo_weight = value) %>%
   mutate(model  = c("binary_delay_model","int_model", "quad_model"),
          loo_weight = loo_weight %>% round(digits = 4)) %>%
   select(model, loo_weight)

 # Combine weights and scores
 model_comparison_exp <- comp_waic_exp %>%select(model,waic, se_waic)%>%
   left_join(weights_waic_exp)%>%
   left_join(comp_loo_exp %>%select(model,looic, se_looic))%>%
   left_join(weights_loo_exp)%>%
   arrange(desc(waic_weight))

 write_csv(model_comparison_exp, "../saves/confirmatory_model_comparison_exploratory.csv")
}

```
