==> analysis/05_supplementary_analysis.Rmd <==
---
title: "ManyPrimates1_Supplementary_Analysis"
output: html_document
---

```{r setup, include=FALSE}
library(ape)
library(tidyverse)
library(readxl)
library(tidyboot)
library(brms)
library(broom)
library(ggthemes)
library(tidybayes)
library(ggridges)
library(ggpubr)
```

```{r}
species_data <- read_xlsx("../data/species_predictors.xlsx", sheet = 1)%>%
  mutate_all(., .funs = tolower)%>%
  select(-contains("Source"), -contains("Comments"))%>%# remove columns with sources and comments
  rename(vocal_repertoire = `vocal_repertoire (# vocalization types)`,
         body_size = `body_size (kg)`)%>%
  mutate_if(is.character, tolower)%>%
#  filter(species_english != "allens_swamp_monkey",
 #        species_english != "hamlyns_monkey") %>%
  mutate(vocal_repertoire = as.numeric(vocal_repertoire),
         group_size = as.numeric(group_size),
         home_range = as.numeric(home_range),
         feeding_budget = as.numeric(feeding_budget),
         day_journey_length= as.numeric(day_journey_length),
         resting_time_percent_activitybudget= as.numeric(resting_time_percent_activitybudget),
         feeding_budget= as.numeric(feeding_budget),
         dietary_breadth= as.numeric(dietary_breadth),
         percent_frugivory= as.numeric(percent_frugivory),
         body_size= as.numeric(body_size),
         home_range_body_site_ratio = home_range/body_size)%>%
  mutate(resting_time_percent_activitybudget = scale(resting_time_percent_activitybudget, center = T, scale = T),
         home_range_body_site_ratio = scale(home_range_body_site_ratio, center = T, scale = T),
         home_range = scale(home_range, center = T, scale = T),
         feeding_budget = scale(feeding_budget, center = T, scale = T),
         dietary_breadth = scale(dietary_breadth, center = T, scale = T),
         vocal_repertoire = scale(vocal_repertoire, center = T, scale = T),
         group_size = scale(group_size, center = T, scale = T),
         day_journey_length = scale(day_journey_length, center = T, scale = T),
         percent_frugivory = scale(percent_frugivory, center = T, scale = T),
         feeding_budget = scale(feeding_budget, center = T, scale =T))

data <- read_csv("../data/merged_data/ManyPrimates_mp1_merged_data.csv")%>%
  left_join(species_data%>%select(species_english, superordinate_group)%>%rename(species = species_english))%>%
  mutate(superordinate_group = factor(superordinate_group, levels = c("ape", "old_world_monkey", "new_world_monkey", "lemur"), labels = c("Hominoidea", "Cercopithecoidea", "Platyrrhini", "Strepsirrhini")))
```

# Relation between age and STM performance

```{r}
age <- data %>%
  filter(condition != "short")%>%
  group_by(superordinate_group, species, subject_site,norm_age)%>%
  summarise(mean = mean(correct))

ggplot(age, aes(x = norm_age, y = mean, col = superordinate_group, pch = superordinate_group))+
  geom_point(alpha = .75)+
  geom_smooth(method = "lm", inherit.aes = F, aes(x = norm_age, y = mean),col = "firebrick", size = 0.5)+
  labs (x = "Normed age", y = "STM performance")+
  scale_color_colorblind(name = "Clade")+
  scale_shape(name = "Clade")+
  theme_minimal()
```
```{r}
ggsave("../figures/age.png", height = 4, width = 7)
```

# Split-half reliability of delayed response task

Here we split the data into odd and even trials to assess split-half reliability of the task. We exclude the short delay trials to get a measure that loads more strongly on short-term memory abilities. The result does not change much if these trials are included as well. 

```{r}
rel <- data %>%
  filter(condition != "short")%>%
  mutate(test_half = ifelse(trial %%2 , "even_trials", "odd_trials"))%>%
  group_by(superordinate_group, species, subject_site, test_half)%>%
  summarise(mean = mean(correct))%>%
  pivot_wider(names_from = test_half, values_from = mean)

ggplot(rel, aes(x = even_trials, y = odd_trials, col = superordinate_group, pch = superordinate_group))+
  geom_abline(intercept = 0, slope = 1, lty = 2, alpha = 0.7, size = 0.5)+
  stat_cor(method = "pearson", aes(x = even_trials, y = odd_trials, label = paste(..r.label..)),cor.coef.name = "r", inherit.aes = F, size = 3)+
  geom_count(alpha = .75)+
  labs (x = "Test-half 1 (even trials)", y = "Test-half 2 (odd trials)")+
  scale_color_colorblind(name = "Clade")+
  scale_shape(name = "Clade")+
  theme_minimal()
```
```{r}
ggsave("../figures/split_half_reliability.png", height = 4, width = 7)
```


# Comparison across sites

We visualize the performance for one species at different data collection sites. We only included sites with at least 5 individuals of a species. 

```{r}
site_1 <- data %>%
  group_by(species, site)%>%
  mutate(subject_per_site = length(unique(subject_site)))%>%
  filter(subject_per_site > 5)%>%
  group_by(species)%>%
  mutate(sites_species = length(unique(site)))%>%
  filter(sites_species > 1)%>%
  mutate(site = paste("Site",as.numeric(factor(site))))%>%
  group_by(species, site, subject_site, condition)%>%
  summarise(mean = mean(correct))

site_2 <- site_1%>%
  group_by(species, site, condition)%>%
  tidyboot_mean(column = mean)

ggplot()+
  geom_hline(yintercept = 1/3, lty = 2)+
  geom_point(data = site_1, aes(x = site, y = mean, col = condition), position = position_jitterdodge(jitter.height = 0., jitter.width = 0.1, dodge.width = 0.5), alpha = .25, width = .05, height = 0)+
  geom_pointrange(data = site_2, aes(x = site, y = mean, col = condition, ymin = ci_lower, ymax = ci_upper, col = condition), position = position_dodge(width = .5))+
  labs(x = "Site", y = "Proportion Correct")+
  scale_color_ptol(name = "Delay")+
  facet_wrap(species~., scales = "free_y")+
  ylim(-0.05, 1.05)+
  theme_few()+
  coord_flip()

```

```{r}
ggsave("../figures/sites.png", height = 4, width = 10)
```

# Correlation between STM performance and predictor variables

Here we correlate STM performance with the numeric predictor variables. STM performance was computed based on the medium and long delay trials. the results do not change much when short delay trials are also included. 

```{r}
pred_cor <- data%>%
  filter(species != "allens_swamp_monkey",
        species != "hamlyns_monkey") %>%
  filter(condition != "short")%>%
  group_by(superordinate_group, species)%>%
  summarise(mean = mean(correct))%>%
  right_join(
    species_data%>%
      filter(species_english != "allens_swamp_monkey",
             species_english != "hamlyns_monkey")%>%
      mutate(superordinate_group = factor(superordinate_group, levels = c("ape", "old_world_monkey", "new_world_monkey", "lemur"), labels = c("Hominoidea", "Cercopithecoidea", "Platyrrhini", "Strepsirrhini")))%>%
      rename(species = species_english)%>%
      select(-color_vision, -diet_diversity,-terrestriality, -life_expectancy, -body_size)%>%
    pivot_longer(cols = c(vocal_repertoire,group_size,home_range,day_journey_length,resting_time_percent_activitybudget,feeding_budget,dietary_breadth,percent_frugivory, home_range_body_site_ratio), 
                 names_to = "predictor", values_to = "value"))


ggplot(pred_cor, aes(x = value, y = mean, col = superordinate_group, pch = superordinate_group))+
  #geom_abline(intercept = 0, slope = 1, lty = 2, alpha = 0.7, size = 0.5)+
  geom_smooth(method = "lm", inherit.aes = F, aes(x = value, y = mean), col = "firebrick", size = 0.5)+
  geom_jitter(alpha = .75)+
  stat_cor(method = "pearson", aes(x = value, y = mean, label = paste(..r.label..)),cor.coef.name = "r", inherit.aes = F, size = 3)+
  labs (y = "STM performance", x = "Predictor (scaled)")+
  facet_wrap(~predictor, nrow = 3, scales = "free_x")+
  scale_color_colorblind(name = "Clade")+
  scale_shape(name = "Clade")+
  ylim(0,1)+
  theme_few()+
  theme(legend.position = "bottom")
```

```{r}
ggsave("../figures/correlation_predictor_stm.png", height = 8, width = 8)
```

# Visualize predictor models and baseline

```{r}
 base_model  <- readRDS("../saves/phylo_base_model.rds")  
 vocal_repertoire_model <- readRDS("../saves/phylo_vocal_repertoire_model.rds") 
 dietary_breadth_model <- readRDS("../saves/phylo_dietary_breadth_model.rds")
```

## Posterior distributions for predictors
```{r}
model_draws <-  bind_rows(
  vocal_repertoire_model%>% posterior_samples(pars = c("b_delay","b_vocal_repertoire","b_delay:vocal_repertoire"))%>%
    rename(delay = b_delay,
           predictor = b_vocal_repertoire,
           interaction = `b_delay:vocal_repertoire`)%>%
    mutate(model= "vocal_repertoire"),
  dietary_breadth_model%>% posterior_samples(pars = c("b_delay","b_dietary_breadth","b_delay:dietary_breadth"))%>%
    rename(delay = b_delay,
           predictor = b_dietary_breadth,
           interaction = `b_delay:dietary_breadth`)%>%
    mutate(model= "dietary_breadth"),
  base_model%>% posterior_samples(pars = c("b_delay"))%>%
    rename(delay = b_delay)%>%
    mutate(model= "baseline")
)%>%
  pivot_longer(names_to = "estimate", values_to = "value", cols = c(delay,predictor,interaction ))

model_draws %>%
  ggplot(.,aes(x = value, y= model, fill = factor(stat(quantile))))+
  geom_vline(xintercept = 0, lty = 2)+
  stat_density_ridges(
    geom = "density_ridges_gradient",
    calc_ecdf = TRUE,
    quantiles = c(0.025, 0.975)
  ) +
  facet_grid(~estimate)+
  theme_minimal() +
  scale_fill_manual(
   values = c("firebrick", "NA", "firebrick")
  )+
    labs(x = "Estimate", y = "Model")+
  guides(fill = F)
```

```{r}
ggsave("../figures/phylo_model_predictors.png", height = 3, width = 7)
```

# Relation between body size and cup distance

```{r}
cup <- data%>%
  filter(condition != "short")%>%
  group_by(superordinate_group, species, subject_site)%>%
  summarise(cup_distance = mean(cup_distance), 
            mean = mean(correct))%>%
  left_join(
    species_data%>%
      mutate(superordinate_group = factor(superordinate_group, levels = c("ape", "old_world_monkey", "new_world_monkey", "lemur"), labels = c("Hominoidea", "Cercopithecoidea", "Platyrrhini", "Strepsirrhini")))%>%
      rename(species = species_english)%>%
      select(species,superordinate_group,body_size))

ggplot(cup, aes(x = cup_distance, y = body_size, col = mean, pch = superordinate_group))+
  stat_cor(method = "pearson", aes(x = cup_distance, y = body_size, label = paste(..r.label..)),cor.coef.name = "r", inherit.aes = F, size = 3)+
  geom_smooth(method = "lm", inherit.aes = F, aes(x = cup_distance, y = body_size), col = "firebrick", size = 0.5)+
  geom_jitter(alpha = .5)+
  labs (x = "Cup distance", y = "Body size")+
  scale_color_viridis_c(name = "Performance in delay trials")+
  scale_shape(name = "Clade")+
  theme_minimal()
```
```{r}
ggsave("../figures/body_size_cup_distance.png", height = 4, width = 8)
```


# Overview plot 

```{r}
o1 <- data %>%
  group_by(superordinate_group, species,condition)%>%
  mutate(n = length(unique(subject_site)))%>%
  ungroup()%>%
  group_by(superordinate_group, species,subject_site,n, condition)%>%
  summarise(mean = mean(correct))

o2 <- o1%>%
  group_by(superordinate_group, species, n, condition)%>%
  tidyboot_mean(column = mean)

ggplot()+
  geom_hline(yintercept = 1/3, lty = 2)+
  geom_count(data = o1, aes(x = species, y = mean, col = condition), position = position_jitterdodge(jitter.height = 0, jitter.width = 0.1, dodge.width = 0.5), alpha = .25, width = .05, height = 0)+
  geom_pointrange(data = o2, aes(x = species, y = mean, col = condition, ymin = ci_lower, ymax = ci_upper, col = condition), position = position_dodge(width = .5))+
  labs(x = "Species", y = "Proportion Correct")+
  scale_color_ptol(name = "Delay")+
  facet_grid(~superordinate_group, scales = "free_x")+
  ylim(-0.05, 1.05)+
  theme_few()+
  theme(axis.text.x=element_text(angle = 45, vjust = 1, hjust = 1))

```

```{r}
ggsave("../figures/overview.png", height = 6, width = 16)
```
